Fix "para" pattern factory default suffix
Unlike paraValue — which only stops at a new line character followed by
a new key + div —, "para" is stopping at the first new line character of
all.

Because of that, it's only capturing lines, and not paragraphs.

Change it so that it stops at a new line character followed by zero or
more spaces or tabs followed by another new line character.