Fix subsequent key => value pairs being included in lineValue result
If there are multiple key => value pairs on the same line, only the
desired key's value will be returned.

A key following the desired one is expected to be composed of a single
word, since the first words of a multi-worded key
will create ambiguity and be treated as part of the desired key's value.

For that reason and to avoid the confusion of alternating between
single- and multi-worded keys, it's best to use only keys composed of
single words when using lineValue.